executor: migrate test-infra to testify for executor_test.testSuite(6|8)
