+++ executor/executor_legacy_test.go
@@ -58,7 +58,6 @@ import (
 	"github.com/pingcap/tidb/store/copr"
 	error2 "github.com/pingcap/tidb/store/driver/error"
 	"github.com/pingcap/tidb/store/mockstore"
-	testkit2 "github.com/pingcap/tidb/testkit"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util"
 	"github.com/pingcap/tidb/util/admin"
@@ -68,7 +67,6 @@ import (
 	"github.com/pingcap/tidb/util/mock"
 	"github.com/pingcap/tidb/util/testkit"
 	"github.com/pingcap/tidb/util/testleak"
-	"github.com/stretchr/testify/require"
 	"github.com/tikv/client-go/v2/oracle"
 	"github.com/tikv/client-go/v2/testutils"
 	"github.com/tikv/client-go/v2/tikv"
@@ -89,8 +87,6 @@ var _ = Suite(&testSuite1{})
 var _ = SerialSuites(&testSuiteWithCliBaseCharset{})
 var _ = Suite(&testSuite2{&baseTestSuite{}})
 var _ = Suite(&testSuite3{&baseTestSuite{}})
-var _ = Suite(&testSuite6{&baseTestSuite{}})
-var _ = Suite(&testSuite8{&baseTestSuite{}})
 var _ = SerialSuites(&testRecoverTable{})
 var _ = SerialSuites(&testClusterTableSuite{})
 var _ = SerialSuites(&testSplitTable{&baseTestSuite{}})
@@ -467,48 +463,6 @@ func (s *testSuiteP2) TestAdminChecksumOfPartitionedTable(c *C) {
 	r.Check(testkit.Rows("test admin_checksum_partition_test 1 5 5"))
 }
 
-type testCase struct {
-	data1       []byte
-	data2       []byte
-	expected    []string
-	restData    []byte
-	expectedMsg string
-}
-
-func checkCases(tests []testCase, ld *executor.LoadDataInfo,
-	t *testing.T, tk *testkit2.TestKit, ctx sessionctx.Context, selectSQL, deleteSQL string) {
-	origin := ld.IgnoreLines
-	for _, tt := range tests {
-		ld.IgnoreLines = origin
-		require.Nil(t, ctx.NewTxn(context.Background()))
-		ctx.GetSessionVars().StmtCtx.DupKeyAsWarning = true
-		ctx.GetSessionVars().StmtCtx.BadNullAsWarning = true
-		ctx.GetSessionVars().StmtCtx.InLoadDataStmt = true
-		ctx.GetSessionVars().StmtCtx.InDeleteStmt = false
-		data, reachLimit, err1 := ld.InsertData(context.Background(), tt.data1, tt.data2)
-		require.NoError(t, err1)
-		require.False(t, reachLimit)
-		err1 = ld.CheckAndInsertOneBatch(context.Background(), ld.GetRows(), ld.GetCurBatchCnt())
-		require.NoError(t, err1)
-		ld.SetMaxRowsInBatch(20000)
-		if tt.restData == nil {
-			require.Len(t, data, 0, "data1:%v, data2:%v, data:%v", string(tt.data1), string(tt.data2), string(data))
-		} else {
-			require.Equal(t, tt.restData, data, "data1:%v, data2:%v, data:%v", string(tt.data1), string(tt.data2), string(data))
-		}
-		ld.SetMessage()
-		require.Equal(t, tt.expectedMsg, tk.Session().LastMessage())
-		ctx.StmtCommit()
-		txn, err := ctx.Txn(true)
-		require.NoError(t, err)
-		err = txn.Commit(context.Background())
-		require.NoError(t, err)
-		r := tk.MustQuery(selectSQL)
-		r.Check(testkit.RowsWithSep("|", tt.expected...))
-		tk.MustExec(deleteSQL)
-	}
-}
-
 func (s *testSuiteP2) TestUnion(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
@@ -765,89 +719,6 @@ func (s *testSuite2) TestUnionLimit(c *C) {
 	}
 	// Cover the code for worker count limit in the union executor.
 	tk.MustQuery("select * from union_limit limit 10")
-}
-
-func (s *testSuite8) TestIndexScan(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (a int unique)")
-	tk.MustExec("insert t values (-1), (2), (3), (5), (6), (7), (8), (9)")
-	result := tk.MustQuery("select a from t where a < 0 or (a >= 2.1 and a < 5.1) or ( a > 5.9 and a <= 7.9) or a > '8.1'")
-	result.Check(testkit.Rows("-1", "3", "5", "6", "7", "9"))
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (a int unique)")
-	tk.MustExec("insert t values (0)")
-	result = tk.MustQuery("select NULL from t ")
-	result.Check(testkit.Rows("<nil>"))
-	// test for double read
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (a int unique, b int)")
-	tk.MustExec("insert t values (5, 0)")
-	tk.MustExec("insert t values (4, 0)")
-	tk.MustExec("insert t values (3, 0)")
-	tk.MustExec("insert t values (2, 0)")
-	tk.MustExec("insert t values (1, 0)")
-	tk.MustExec("insert t values (0, 0)")
-	result = tk.MustQuery("select * from t order by a limit 3")
-	result.Check(testkit.Rows("0 0", "1 0", "2 0"))
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (a int unique, b int)")
-	tk.MustExec("insert t values (0, 1)")
-	tk.MustExec("insert t values (1, 2)")
-	tk.MustExec("insert t values (2, 1)")
-	tk.MustExec("insert t values (3, 2)")
-	tk.MustExec("insert t values (4, 1)")
-	tk.MustExec("insert t values (5, 2)")
-	result = tk.MustQuery("select * from t where a < 5 and b = 1 limit 2")
-	result.Check(testkit.Rows("0 1", "2 1"))
-	tk.MustExec("drop table if exists tab1")
-	tk.MustExec("CREATE TABLE tab1(pk INTEGER PRIMARY KEY, col0 INTEGER, col1 FLOAT, col3 INTEGER, col4 FLOAT)")
-	tk.MustExec("CREATE INDEX idx_tab1_0 on tab1 (col0)")
-	tk.MustExec("CREATE INDEX idx_tab1_1 on tab1 (col1)")
-	tk.MustExec("CREATE INDEX idx_tab1_3 on tab1 (col3)")
-	tk.MustExec("CREATE INDEX idx_tab1_4 on tab1 (col4)")
-	tk.MustExec("INSERT INTO tab1 VALUES(1,37,20.85,30,10.69)")
-	result = tk.MustQuery("SELECT pk FROM tab1 WHERE ((col3 <= 6 OR col3 < 29 AND (col0 < 41)) OR col3 > 42) AND col1 >= 96.1 AND col3 = 30 AND col3 > 17 AND (col0 BETWEEN 36 AND 42)")
-	result.Check(testkit.Rows())
-	tk.MustExec("drop table if exists tab1")
-	tk.MustExec("CREATE TABLE tab1(pk INTEGER PRIMARY KEY, a INTEGER, b INTEGER)")
-	tk.MustExec("CREATE INDEX idx_tab1_0 on tab1 (a)")
-	tk.MustExec("INSERT INTO tab1 VALUES(1,1,1)")
-	tk.MustExec("INSERT INTO tab1 VALUES(2,2,1)")
-	tk.MustExec("INSERT INTO tab1 VALUES(3,1,2)")
-	tk.MustExec("INSERT INTO tab1 VALUES(4,2,2)")
-	result = tk.MustQuery("SELECT * FROM tab1 WHERE pk <= 3 AND a = 1")
-	result.Check(testkit.Rows("1 1 1", "3 1 2"))
-	result = tk.MustQuery("SELECT * FROM tab1 WHERE pk <= 4 AND a = 1 AND b = 2")
-	result.Check(testkit.Rows("3 1 2"))
-	tk.MustExec("CREATE INDEX idx_tab1_1 on tab1 (b, a)")
-	result = tk.MustQuery("SELECT pk FROM tab1 WHERE b > 1")
-	result.Check(testkit.Rows("3", "4"))
-
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("CREATE TABLE t (a varchar(3), index(a))")
-	tk.MustExec("insert t values('aaa'), ('aab')")
-	result = tk.MustQuery("select * from t where a >= 'aaaa' and a < 'aabb'")
-	result.Check(testkit.Rows("aab"))
-
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("CREATE TABLE t (a int primary key, b int, c int, index(c))")
-	tk.MustExec("insert t values(1, 1, 1), (2, 2, 2), (4, 4, 4), (3, 3, 3), (5, 5, 5)")
-	// Test for double read and top n.
-	result = tk.MustQuery("select a from t where c >= 2 order by b desc limit 1")
-	result.Check(testkit.Rows("5"))
-
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a varchar(50) primary key, b int, c int, index idx(b))")
-	tk.MustExec("insert into t values('aa', 1, 1)")
-	tk.MustQuery("select * from t use index(idx) where a > 'a'").Check(testkit.Rows("aa 1 1"))
-
-	// fix issue9636
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("CREATE TABLE `t` (a int, KEY (a))")
-	result = tk.MustQuery(`SELECT * FROM (SELECT * FROM (SELECT a as d FROM t WHERE a IN ('100')) AS x WHERE x.d < "123" ) tmp_count`)
-	result.Check(testkit.Rows())
 }
 
 func (s *testSuiteP2) TestToPBExpr(c *C) {
@@ -2213,84 +2084,6 @@ func (s *testSuite3) TestSortLeftJoinWithNullColumnInRightChildPanic(c *C) {
 		Check(testkit.Rows("<nil>"))
 }
 
-func (s *testSuite6) TestUpdateJoin(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t1, t2, t3, t4, t5, t6, t7")
-	tk.MustExec("create table t1(k int, v int)")
-	tk.MustExec("create table t2(k int, v int)")
-	tk.MustExec("create table t3(id int auto_increment, k int, v int, primary key(id))")
-	tk.MustExec("create table t4(k int, v int)")
-	tk.MustExec("create table t5(v int, k int, primary key(k))")
-	tk.MustExec("insert into t1 values (1, 1)")
-	tk.MustExec("insert into t4 values (3, 3)")
-	tk.MustExec("create table t6 (id int, v longtext)")
-	tk.MustExec("create table t7 (x int, id int, v longtext, primary key(id))")
-
-	// test the normal case that update one row for a single table.
-	tk.MustExec("update t1 set v = 0 where k = 1")
-	tk.MustQuery("select k, v from t1 where k = 1").Check(testkit.Rows("1 0"))
-
-	// test the case that the table with auto_increment or none-null columns as the right table of left join.
-	tk.MustExec("update t1 left join t3 on t1.k = t3.k set t1.v = 1")
-	tk.MustQuery("select k, v from t1").Check(testkit.Rows("1 1"))
-	tk.MustQuery("select id, k, v from t3").Check(testkit.Rows())
-
-	// test left join and the case that the right table has no matching record but has updated the right table columns.
-	tk.MustExec("update t1 left join t2 on t1.k = t2.k set t1.v = t2.v, t2.v = 3")
-	tk.MustQuery("select k, v from t1").Check(testkit.Rows("1 <nil>"))
-	tk.MustQuery("select k, v from t2").Check(testkit.Rows())
-
-	// test the case that the update operation in the left table references data in the right table while data of the right table columns is modified.
-	tk.MustExec("update t1 left join t2 on t1.k = t2.k set t2.v = 3, t1.v = t2.v")
-	tk.MustQuery("select k, v from t1").Check(testkit.Rows("1 <nil>"))
-	tk.MustQuery("select k, v from t2").Check(testkit.Rows())
-
-	// test right join and the case that the left table has no matching record but has updated the left table columns.
-	tk.MustExec("update t2 right join t1 on t2.k = t1.k set t2.v = 4, t1.v = 0")
-	tk.MustQuery("select k, v from t1").Check(testkit.Rows("1 0"))
-	tk.MustQuery("select k, v from t2").Check(testkit.Rows())
-
-	// test the case of right join and left join at the same time.
-	tk.MustExec("update t1 left join t2 on t1.k = t2.k right join t4 on t4.k = t2.k set t1.v = 4, t2.v = 4, t4.v = 4")
-	tk.MustQuery("select k, v from t1").Check(testkit.Rows("1 0"))
-	tk.MustQuery("select k, v from t2").Check(testkit.Rows())
-	tk.MustQuery("select k, v from t4").Check(testkit.Rows("3 4"))
-
-	// test normal left join and the case that the right table has matching rows.
-	tk.MustExec("insert t2 values (1, 10)")
-	tk.MustExec("update t1 left join t2 on t1.k = t2.k set t2.v = 11")
-	tk.MustQuery("select k, v from t2").Check(testkit.Rows("1 11"))
-
-	// test the case of continuously joining the same table and updating the unmatching records.
-	tk.MustExec("update t1 t11 left join t2 on t11.k = t2.k left join t1 t12 on t2.v = t12.k set t12.v = 233, t11.v = 111")
-	tk.MustQuery("select k, v from t1").Check(testkit.Rows("1 111"))
-	tk.MustQuery("select k, v from t2").Check(testkit.Rows("1 11"))
-
-	// test the left join case that the left table has records but all records are null.
-	tk.MustExec("delete from t1")
-	tk.MustExec("delete from t2")
-	tk.MustExec("insert into t1 values (null, null)")
-	tk.MustExec("update t1 left join t2 on t1.k = t2.k set t1.v = 1")
-	tk.MustQuery("select k, v from t1").Check(testkit.Rows("<nil> 1"))
-
-	// test the case that the right table of left join has an primary key.
-	tk.MustExec("insert t5 values(0, 0)")
-	tk.MustExec("update t1 left join t5 on t1.k = t5.k set t1.v = 2")
-	tk.MustQuery("select k, v from t1").Check(testkit.Rows("<nil> 2"))
-	tk.MustQuery("select k, v from t5").Check(testkit.Rows("0 0"))
-
-	tk.MustExec("insert into t6 values (1, NULL)")
-	tk.MustExec("insert into t7 values (5, 1, 'a')")
-	tk.MustExec("update t6, t7 set t6.v = t7.v where t6.id = t7.id and t7.x = 5")
-	tk.MustQuery("select v from t6").Check(testkit.Rows("a"))
-
-	tk.MustExec("drop table if exists t1, t2")
-	tk.MustExec("create table t1(id int primary key, v int, gv int GENERATED ALWAYS AS (v * 2) STORED)")
-	tk.MustExec("create table t2(id int, v int)")
-	tk.MustExec("update t1 tt1 inner join (select count(t1.id) a, t1.id from t1 left join t2 on t1.id = t2.id group by t1.id) x on tt1.id = x.id set tt1.v = tt1.v + x.a")
-}
-
 func (s *testSuite3) TestMaxOneRow(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec(`use test`)
@@ -2510,46 +2303,6 @@ type testSuite3 struct {
 }
 
 func (s *testSuite3) TearDownTest(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	r := tk.MustQuery("show full tables")
-	for _, tb := range r.Rows() {
-		tableName := tb[0]
-		if tb[1] == "VIEW" {
-			tk.MustExec(fmt.Sprintf("drop view %v", tableName))
-		} else if tb[1] == "SEQUENCE" {
-			tk.MustExec(fmt.Sprintf("drop sequence %v", tableName))
-		} else {
-			tk.MustExec(fmt.Sprintf("drop table %v", tableName))
-		}
-	}
-}
-
-type testSuite6 struct {
-	*baseTestSuite
-}
-
-func (s *testSuite6) TearDownTest(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	r := tk.MustQuery("show full tables")
-	for _, tb := range r.Rows() {
-		tableName := tb[0]
-		if tb[1] == "VIEW" {
-			tk.MustExec(fmt.Sprintf("drop view %v", tableName))
-		} else if tb[1] == "SEQUENCE" {
-			tk.MustExec(fmt.Sprintf("drop sequence %v", tableName))
-		} else {
-			tk.MustExec(fmt.Sprintf("drop table %v", tableName))
-		}
-	}
-}
-
-type testSuite8 struct {
-	*baseTestSuite
-}
-
-func (s *testSuite8) TearDownTest(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	r := tk.MustQuery("show full tables")
